+++ package.json
@@ -47,6 +47,7 @@
     "react-icons": "^4.1.0",
     "react-linkify": "^1.0.0-alpha",
     "react-loading-skeleton": "^2.1.1",
+    "react-masonry-component": "^6.2.1",
     "react-modal": "^3.12.1",
     "react-robust-accordion": "^0.5.2",
     "react-select": "^4.1.0",
